Support more response codes than just 200 OK.
The Rundeck API returns 201 Created for newly-created objects, and it
returns 204 No Content when deleting objects. To support both of these
use-cases we accept all 2xx responses as successful, and then we allow
the caller to decide (by passing nil as the "result") to not require
a response payload.

We assume that only 200 and 201 responses return response payloads. If
a different response code is returned but the client wanted a body then
that is considered an error.