Empty out RawConfigItems when we "inflate" a project.
When we "deflate" a project we combine the two config representations
together, so for symmetry we must empty the RawConfigItems such that
we aren't effectively duplicating each of the settings from a request
when we send the same object back in a response.

This essentially makes the RawConfigItems member unusable by callers,
but I'm pretty comfortable with that since it's a pretty awful interface
anyway and I'd make it private if only the XML library wouldn't require
all target members to be public.